adasdadda

+++ source/network_server.c
@@ -202,10 +202,10 @@ int network_main_loop(int listening_socket, struct table_t *table)
  */
 MessageT *network_receive(int client_socket)
 {
-    uint32_t size; // Variavel de apoio para armazenar tmnh da mensagem
+    short size; // Variavel de apoio para armazenar tmnh da mensagem
 
     // Receber o tamanho da mensagem
-    if (read_all(client_socket, &size, sizeof(uint32_t)) <= 0)
+    if (read_all(client_socket, &size, sizeof(short)) <= 0)
     {
         printf("Erro a receber!\n");
         return NULL;
